+++ src/modules/plugins/plugins.module.ts
@@ -5,8 +5,10 @@ import { Module } from '@nestjs/common'
 import { PassportModule } from '@nestjs/passport'
 
 import { ConfigModule } from '../../core/config/config.module.js'
+import { HomebridgeIpcModule } from '../../core/homebridge-ipc/homebridge-ipc.module.js'
 import { LoggerModule } from '../../core/logger/logger.module.js'
 import { NodePtyModule } from '../../core/node-pty/node-pty.module.js'
+import { ChildBridgesModule } from '../child-bridges/child-bridges.module.js'
 import { PluginsController } from './plugins.controller.js'
 import { PluginsGateway } from './plugins.gateway.js'
 import { PluginsService } from './plugins.service.js'
@@ -24,6 +26,8 @@ import { PluginsService } from './plugins.service.js'
     NodePtyModule,
     ConfigModule,
     LoggerModule,
+    HomebridgeIpcModule,
+    ChildBridgesModule,
   ],
   providers: [
     PluginsService,